pyproject.toml
[project]
name = "ptt-crawler-distributed"
version = "1.0.0"
description = "分散式 PTT 爬蟲系統 - 使用 Docker + Celery + RabbitMQ + MySQL"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4>=4.9.3",
    "fake-useragent>=1.4.0",
    "pandas>=1.3.0",
    "requests>=2.25.1",
    "celery==5.5.0",
    "sqlalchemy==1.4.54",
    "pymysql==1.1.1",
    "python-dotenv>=0.19.0",
    "cryptography>=3.4.8",
]

[dependency-groups]
dev = [
    "black>=22.0.0",
    "flake8>=4.0.0",
    "pytest>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["crawler"]

[tool.uv]
dev-dependencies = [
    "black>=22.0.0",
    "flake8>=4.0.0", 
    "pytest>=6.0.0",
]
